requirements.txt
pandas
numpy
pytest
langchain-core
langchain-google-genai
python-dotenv
sentence-transformers
